Remove std::auto_ptr
`std::auto_ptr` should never be used.
It's bad enough that it got removed from the standard in C++17.
In this case there wasn't a need for a smart pointer at all since the object only needed to be constructed to be passed as a non-owning reference to another function.